pyproject.toml
[project]
name = "material-ai"
version = "1.1.18"
description = "Material AI Agent with FastAPI server"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "deprecated==1.2.18",
    "google-adk==1.23.0",
    "psutil>=7.1.0",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.3",
    "sqlalchemy-bigquery>=1.16.0",
    "starlette[full]>=0.50.0",
]

[project.urls]
Repository = "https://github.com/muralimanoharv/material-ai/tree/v1.1.18" # x-release-please-version

[project.optional-dependencies]
dev = [
    "black>=25.9.0",
    "build>=1.3.0",
    "debugpy>=1.8.15",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
    "twine>=6.2.0",
]

[tool.black]
# Set the maximum line length. Black defaults to 88.
line-length = 88 
# Files to skip over
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | \.vscode
  | dist
)/
'''

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.pytest.ini_options]
pythonpath = ["src"]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true
